[delta_q] clean up and document ΔQSD analysis
