+++ topology-checker/src/analysis/mod.rs
@@ -1,7 +1,9 @@
 pub mod basic;
+pub mod delta_q;
 pub mod geographic;
 pub mod stake;
 
 pub use basic::*;
+pub use delta_q::*;
 pub use geographic::*;
 pub use stake::*;
